+++ @navikt/internal/react/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @navikt/ds-react-internal
 
+## 2.9.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @navikt/ds-react@2.9.0
+  - @navikt/ds-icons@2.9.0
+
 ## 2.8.16
 
 ### Patch Changes